+++ src/commonMain/kotlin/io/github/kotlinmania/libc/unix/linuxlike/linux/arch/sparc/Mod.kt
@@ -220,4 +220,4 @@ public const val RLIMIT_RTPRIO: RlimitResourceT = 14
 public const val RLIMIT_RTTIME: RlimitResourceT = 15
 public const val RLIM_NLIMITS: RlimitResourceT = 16
 public const val RLIMIT_NLIMITS: RlimitResourceT = RLIM_NLIMITS
-public val RLIM_INFINITY: RlimT = 0.inv()
+public val RLIM_INFINITY: RlimT = 0.inv().toULong()
